+++ src/app/chat/(desktop)/index.tsx
@@ -1,11 +1,9 @@
 'use client';
 
-import dynamic from 'next/dynamic';
-import { FC, memo } from 'react';
+import { memo } from 'react';
 import { Flexbox } from 'react-layout-kit';
 
 import ClientResponsiveContent from '@/components/client/ClientResponsiveContent';
-import MobileSwitchLoading from '@/features/MobileSwitchLoading';
 
 import PageTitle from '../features/PageTitle';
 import ChatHeader from './features/ChatHeader';
@@ -23,9 +21,4 @@ const Desktop = memo(() => (
   </>
 ));
 
-const Mobile = dynamic(() => import('../(mobile)'), {
-  loading: MobileSwitchLoading,
-  ssr: false,
-}) as FC;
-
-export default ClientResponsiveContent({ Desktop, Mobile });
+export default ClientResponsiveContent({ Desktop, Mobile: () => import('../(mobile)') });
