+++ src/app/chat/settings/(desktop)/index.tsx
@@ -1,10 +1,8 @@
 'use client';
 
-import dynamic from 'next/dynamic';
-import { FC, memo } from 'react';
+import { memo } from 'react';
 
 import ClientResponsiveContent from '@/components/client/ClientResponsiveContent';
-import MobileSwitchLoading from '@/features/MobileSwitchLoading';
 
 import EditPage from '../features/EditPage';
 import Layout from './layout';
@@ -15,9 +13,4 @@ const Desktop = memo(() => (
   </Layout>
 ));
 
-const Mobile = dynamic(() => import('../(mobile)'), {
-  loading: MobileSwitchLoading,
-  ssr: false,
-}) as FC;
-
-export default ClientResponsiveContent({ Desktop, Mobile });
+export default ClientResponsiveContent({ Desktop, Mobile: () => import('../(mobile)') });
