+++ src/app/market/(desktop)/index.tsx
@@ -1,11 +1,9 @@
 'use client';
 
 import { SpotlightCard, SpotlightCardProps } from '@lobehub/ui';
-import dynamic from 'next/dynamic';
 import { FC, memo } from 'react';
 
 import ClientResponsiveContent from '@/components/client/ClientResponsiveContent';
-import MobileSwitchLoading from '@/features/MobileSwitchLoading';
 
 import AgentCard from '../features/AgentCard';
 import Index from '../index';
@@ -17,9 +15,4 @@ const Desktop = memo(() => (
   </>
 ));
 
-const Mobile = dynamic(() => import('../(mobile)'), {
-  loading: MobileSwitchLoading,
-  ssr: false,
-}) as FC;
-
-export default ClientResponsiveContent({ Desktop, Mobile });
+export default ClientResponsiveContent({ Desktop, Mobile: () => import('../(mobile)') });
