+++ src/app/market/(mobile)/index.tsx
@@ -1,16 +1,21 @@
 'use client';
 
+import dynamic from 'next/dynamic';
 import { memo } from 'react';
+import { Flexbox } from 'react-layout-kit';
 
-import AgentCard from '@/app/market/features/AgentCard';
-
+import AgentCard from '../features/AgentCard';
 import Index from '../index';
 import CardRender from './features/AgentCard';
-import Layout from './layout.mobile';
+
+const DetailModal = dynamic(() => import('./features/AgentDetail'), { ssr: false });
 
 export default memo(() => (
-  <Layout>
-    <Index />
-    <AgentCard CardRender={CardRender} mobile />
-  </Layout>
+  <>
+    <Flexbox flex={1} gap={16} style={{ padding: 16 }}>
+      <Index />
+      <AgentCard CardRender={CardRender} mobile />{' '}
+    </Flexbox>
+    <DetailModal />
+  </>
 ));
