+++ src/app/market/_layout/Desktop/index.tsx
@@ -3,13 +3,12 @@
 import { GridBackground } from '@lobehub/ui';
 import { createStyles } from 'antd-style';
 import dynamic from 'next/dynamic';
-import { FC, PropsWithChildren, memo } from 'react';
+import { PropsWithChildren, memo } from 'react';
 import { Center, Flexbox } from 'react-layout-kit';
 
 import SafeSpacing from '@/components/SafeSpacing';
-import ClientLayout from '@/components/client/ClientLayout';
+import ClientResponsiveLayout from '@/components/client/ClientResponsiveLayout';
 import { MAX_WIDTH } from '@/const/layoutTokens';
-import MobileSwitchLoading from '@/features/MobileSwitchLoading';
 
 import Header from './Header';
 
@@ -57,9 +56,4 @@ const Desktop = memo<PropsWithChildren>(({ children }) => {
   );
 });
 
-const Mobile = dynamic(() => import('../Mobile'), {
-  loading: MobileSwitchLoading,
-  ssr: false,
-}) as FC<PropsWithChildren>;
-
-export default ClientLayout({ Desktop, Mobile });
+export default ClientResponsiveLayout({ Desktop, Mobile: () => import('../Mobile') });
