+++ src/app/welcome/(desktop)/index.tsx
@@ -1,10 +1,8 @@
 'use client';
 
-import dynamic from 'next/dynamic';
-import { FC, memo } from 'react';
+import { memo } from 'react';
 
 import ClientResponsiveContent from '@/components/client/ClientResponsiveContent';
-import MobileSwitchLoading from '@/features/MobileSwitchLoading';
 
 import Footer from './features/Footer';
 import Showcase from './features/Showcase';
@@ -16,9 +14,4 @@ const Desktop = memo(() => (
   </>
 ));
 
-const Mobile = dynamic(() => import('../(mobile)'), {
-  loading: MobileSwitchLoading,
-  ssr: false,
-}) as FC;
-
-export default ClientResponsiveContent({ Desktop, Mobile });
+export default ClientResponsiveContent({ Desktop, Mobile: () => import('../(mobile)') });
