+++ src/app/welcome/_layout/Desktop.tsx
@@ -1,12 +1,10 @@
 'use client';
 
 import { Logo } from '@lobehub/ui';
-import dynamic from 'next/dynamic';
-import { FC, PropsWithChildren, memo } from 'react';
+import { PropsWithChildren, memo } from 'react';
 import { Center, Flexbox } from 'react-layout-kit';
 
-import ClientLayout from '@/components/client/ClientLayout';
-import MobileSwitchLoading from '@/features/MobileSwitchLoading';
+import ClientResponsiveLayout from '@/components/client/ClientResponsiveLayout';
 
 import { useStyles } from '../features/Banner/style';
 
@@ -28,9 +26,4 @@ const Desktop = memo<PropsWithChildren>(({ children }) => {
   );
 });
 
-const Mobile = dynamic(() => import('./Mobile'), {
-  loading: MobileSwitchLoading,
-  ssr: false,
-}) as FC<PropsWithChildren>;
-
-export default ClientLayout({ Desktop, Mobile });
+export default ClientResponsiveLayout({ Desktop, Mobile: () => import('./Mobile') });
