+++ src/components/client/ClientResponsiveContent/index.tsx
@@ -1,19 +1,27 @@
 'use client';
 
+import { Loader } from 'next/dist/shared/lib/dynamic';
+import dynamic from 'next/dynamic';
 import { FC, memo } from 'react';
 
+import MobileSwitchLoading from '@/features/MobileSwitchLoading';
 import { useIsMobile } from '@/hooks/useIsMobile';
 
 interface ClientResponsiveContentProps {
   Desktop: FC;
-  Mobile: FC;
+  Mobile: Loader;
 }
 
 const ClientResponsiveContent = ({ Mobile, Desktop }: ClientResponsiveContentProps) => {
+  const MobileComponent = dynamic(Mobile, {
+    loading: MobileSwitchLoading,
+    ssr: false,
+  });
+
   const Content = memo(() => {
     const mobile = useIsMobile();
 
-    return mobile ? <Mobile /> : <Desktop />;
+    return mobile ? <MobileComponent /> : <Desktop />;
   });
 
   Content.displayName = 'ClientResponsiveContent';
